Refactor: Pep8ified a bit, and made use of pathlib (#8)
* Refactor: Pep8ified a bit, and made use of pathlib

* Bugfix: Minor fixes, not yet working

* Bugfix: All tests run

* Refactor: Removed dead lines of code
