+++ overtime_calculator/src/__init__.py
@@ -4,6 +4,21 @@ from datetime import timedelta
 import logging
 
 
+default_parse_fmt = "%d-%m-%Y %H:%M:%S"
+
+
+def get_secret():
+    return 'sflkjsdjkfd'
+
+
+def token_verify(token):
+    secret = get_secret()
+    try:
+        return jwt.decode(token, secret, algorithm='HS256')
+    except jwt.DecodeError:
+        return False
+
+
 def log_function_entry_and_exit(decorated_function):
     '''
     Function decorator logging time spent.
