+++ overtime_calculator/src/api.py
@@ -1,7 +1,8 @@
 import hug
+
 from overtime_calculator.src import auth
+
 
 @hug.extend_api()
 def auth_api():
     return [auth]
-
