+++ overtime_calculator/tests/auth_test.py
@@ -3,10 +3,12 @@ import shutil
 import pytest
 
 from overtime_calculator.src import api
+from overtime_calculator.src.auth import get_user_folder
+
 
 def test_register():
     response = hug.test.post(api, '/register', {'username': 'test_1', 'password': 'test_1'})
-    assert response.data['status'] == 'ok'
+    assert response.data == {'status': 'ok'}
 
 def test_signin():
     response = hug.test.post(api, '/signin', {'username': 'test_1', 'password': 'test_1'})
@@ -14,5 +16,6 @@ def test_signin():
     assert response.data['token'] is not None
 
 def teardown_module():
-    shutil.rmtree('Users/test_1', ignore_errors=False, onerror=None)
+    user_folder = get_user_folder('test1')
+    shutil.rmtree(str(user_folder), ignore_errors=False)
